Refactor: separate code input component
