+++ client/signup/config/step-components.js
@@ -63,7 +63,7 @@ const stepNameToModuleName = {
 	'p2-get-started': 'p2-get-started',
 	'p2-confirm-email': 'p2-confirm-email',
 	'p2-complete-profile': 'p2-complete-profile',
-	'p2-join-workspaces': 'p2-join-workspaces',
+	'p2-join-workspace': 'p2-join-workspace',
 	'plans-business-monthly': 'plans',
 	'plans-ecommerce-monthly': 'plans',
 	'plans-personal-monthly': 'plans',
