+++ client/signup/steps/p2-join-workspace/index.jsx
@@ -1,0 +1,228 @@
+import { Button } from '@wordpress/components';
+import { useState, useEffect, createInterpolateElement } from '@wordpress/element';
+import { __, _n, sprintf } from '@wordpress/i18n';
+import { Icon, chevronRight } from '@wordpress/icons';
+import { useCallback } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import request from 'wpcom-proxy-request';
+import P2StepWrapper from 'calypso/signup/p2-step-wrapper';
+import { fetchCurrentUser } from 'calypso/state/current-user/actions';
+import { getCurrentUserEmail } from 'calypso/state/current-user/selectors';
+import './style.scss';
+import P2JoinWorkspaceCodeInput from './code-input';
+
+function P2JoinWorkspace( { flowName, goToNextStep, positionInFlow, stepName, submitSignupStep } ) {
+	const dispatch = useDispatch();
+	const userEmail = useSelector( getCurrentUserEmail );
+	if ( ! userEmail ) {
+		dispatch( fetchCurrentUser() );
+	}
+
+	const [ eligibleWorkspaces, setEligibleWorkspaces ] = useState( [] );
+	const [ workspaceStatus, setWorkspaceStatus ] = useState( {
+		requesting: null,
+		requested: null,
+		joined: [],
+	} );
+
+	const fetchList = useCallback( async () => {
+		if ( ! userEmail ) {
+			return;
+		}
+
+		const workspaceList = await request( {
+			path: '/p2/preapproved-joining/list-workspaces',
+			apiNamespace: 'wpcom/v2',
+			global: true,
+		} );
+		setEligibleWorkspaces( workspaceList );
+	}, [ userEmail ] );
+
+	useEffect( () => {
+		fetchList();
+	}, [ fetchList ] );
+
+	const handleJoinWorkspaceClick = async ( { id, name } ) => {
+		// Remember which workspace is being requested, for more accurate loading feedback.
+		setWorkspaceStatus( {
+			...workspaceStatus,
+			requesting: id,
+		} );
+
+		const response = await request( {
+			path: '/p2/preapproved-joining/request-code',
+			apiNamespace: 'wpcom/v2',
+			global: true,
+			method: 'POST',
+			body: {
+				hub_id: parseInt( id ),
+			},
+		} );
+
+		//Test only, REMOVE ME!
+		// const response = {
+		// 	success: true,
+		// 	hub_id: id,
+		// };
+
+		if ( response.success ) {
+			setWorkspaceStatus( {
+				...workspaceStatus,
+				requesting: null,
+				requested: { id: parseInt( response.hub_id ), name },
+			} );
+		}
+	};
+
+	const handleCreateWorkspaceClick = () => {
+		submitSignupStep( {
+			stepName,
+		} );
+
+		goToNextStep();
+	};
+
+	const renderWorkspaceActionButton = ( workspaceId, workspaceName ) => {
+		if ( ! workspaceStatus.joined.includes( workspaceId ) ) {
+			const isBusy = workspaceStatus.requesting === workspaceId;
+			return (
+				<Button
+					className="p2-join-workspace__action-join"
+					onClick={ () => {
+						handleJoinWorkspaceClick( {
+							id: workspaceId,
+							name: workspaceName,
+						} );
+					} }
+					isBusy={ isBusy }
+				>
+					{ ! isBusy && __( 'Join' ) }
+				</Button>
+			);
+		}
+
+		return (
+			<Button className="p2-join-workspace__action-open-workspace">
+				<Icon icon={ chevronRight } />
+			</Button>
+		);
+	};
+
+	const renderEligibleWorkspacesList = () => {
+		return (
+			<div className="p2-join-workspace__workspace-list">
+				{ eligibleWorkspaces.map( ( workspace ) => (
+					<div className="p2-join-workspace__workspace" key={ workspace.id }>
+						<div className="p2-join-workspace__workspace-icon">
+							{ workspace.name && workspace.name.charAt( 0 ) }
+						</div>
+						<div className="p2-join-workspace__workspace-name">
+							<a
+								className="p2-join-workspace__workspace-link"
+								href={ workspace.site_url }
+								target="_blank"
+								rel="noreferrer"
+							>
+								{ workspace.name }
+							</a>
+						</div>
+						<div className="p2-join-workspace__workspace-description">
+							<span>
+								{ sprintf(
+									/* translators: %(userCount)d is a number */
+									_n( '%(userCount)d user', '%(userCount)d users', workspace.user_count ),
+									{
+										userCount: workspace.user_count,
+									}
+								) }
+							</span>
+							<span>
+								{ sprintf(
+									/* translators: %(siteCount)d is a number */
+									_n( '%(siteCount)d P2', '%(siteCount)d P2s', workspace.site_count ),
+									{
+										siteCount: workspace.site_count,
+									}
+								) }
+							</span>
+						</div>
+						<div className="p2-join-workspace__action">
+							{ renderWorkspaceActionButton( parseInt( workspace.id ), workspace.name ) }
+						</div>
+					</div>
+				) ) }
+			</div>
+		);
+	};
+
+	const renderCreateWorkspaceSection = () => {
+		return (
+			<div className="p2-join-workspace__create-workspace">
+				<div className="p2-join-workspace__create-workspace-header">
+					{ __( 'Want to create a new workspace?' ) }
+				</div>
+				<div className="p2-join-workspace__create-workspace-subheader">
+					{ __( 'Get your team on P2 — for free.' ) }
+				</div>
+				<div className="p2-join-workspace__create-workspace-action">
+					<Button onClick={ handleCreateWorkspaceClick }>{ __( 'Create a new workspace' ) }</Button>
+				</div>
+			</div>
+		);
+	};
+
+	const getHeaderText = () => {
+		if ( workspaceStatus.requested?.name ) {
+			return createInterpolateElement(
+				sprintf(
+					/* translators: %s is the site name */
+					__( 'Enter the code to join <br />%s' ),
+					workspaceStatus.requested.name
+				),
+				{ br: <br /> }
+			);
+		}
+
+		return __( 'Welcome to P2!' );
+	};
+
+	const getSubHeaderText = () => {
+		if ( workspaceStatus.requested?.id ) {
+			return createInterpolateElement(
+				sprintf(
+					/* translators: %s is the email address */
+					__( "We've sent an email with a code to <br /><strong>%s</strong>" ),
+					userEmail
+				),
+				{ br: <br />, strong: <strong /> }
+			);
+		}
+
+		return __( 'A better way of working.' );
+	};
+
+	return (
+		userEmail && (
+			<P2StepWrapper
+				flowName={ flowName }
+				stepName={ stepName }
+				positionInFlow={ positionInFlow }
+				headerText={ getHeaderText() }
+				subHeaderText={ getSubHeaderText() }
+			>
+				<div className="p2-join-workspace">
+					{ workspaceStatus.requested && (
+						<P2JoinWorkspaceCodeInput
+							workspaceStatus={ workspaceStatus }
+							setWorkspaceStatus={ setWorkspaceStatus }
+						/>
+					) }
+					{ ! workspaceStatus.requested && renderEligibleWorkspacesList() }
+					{ ! workspaceStatus.requested && renderCreateWorkspaceSection() }
+				</div>
+			</P2StepWrapper>
+		)
+	);
+}
+
+export default P2JoinWorkspace;
