+++ client/signup/steps/p2-join-workspace/style.scss
@@ -1,12 +1,12 @@
-.is-p-2-join-workspaces .p2-step-wrapper {
+.is-p-2-join-workspace .p2-step-wrapper {
 	max-width: 620px;
 }
 
-.p2-join-workspaces__workspaces-list {
+.p2-join-workspace__workspace-list {
 	margin-bottom: 2.5em;
 }
 
-.p2-join-workspaces__workspace {
+.p2-join-workspace__workspace {
 	display: grid;
 	grid-template-columns: min-content auto min-content;
 	grid-column-gap: 12px;
@@ -18,7 +18,7 @@
 	}
 }
 
-.p2-join-workspaces__workspace-icon {
+.p2-join-workspace__workspace-icon {
 	background: var( --p2-color-link );
 	grid-area: 1 / 1 / span 2;
 	color: var( --p2-color-white );
@@ -32,10 +32,10 @@
 	border-radius: 2px;
 }
 
-.p2-join-workspaces__workspace-name {
+.p2-join-workspace__workspace-name {
 	font-weight: 600;
 
-	.p2-join-workspaces__workspace-link {
+	.p2-join-workspace__workspace-link {
 		color: var( --p2-color-text );
 	
 		&:hover {
@@ -44,7 +44,7 @@
 	}
 }
 
-.p2-join-workspaces__workspace-description {
+.p2-join-workspace__workspace-description {
 	font-size: 0.75rem;
 	color: var( --p2-color-border-medium );
 	grid-area: 2 / 2;
@@ -61,14 +61,14 @@
 	}
 }
 
-.p2-join-workspaces__action {
+.p2-join-workspace__action {
 	grid-area: 1 / 3 / span 2;
 	align-items: center;
 	display: flex;
 	justify-content: center;
 }
 
-.p2-join-workspaces__action-join {
+.p2-join-workspace__action-join {
 	background: var( --p2-color-background-light );
 	border-radius: 18px; /* stylelint-disable-line scales/radii */
 	padding: 0 1.5rem;
@@ -79,7 +79,7 @@
 	}
 }
 
-.p2-join-workspaces__create-workspace {
+.p2-join-workspace__create-workspace {
 	background: var( --p2-color-background-light );
 	padding: 1.2em;
 	display: grid;
@@ -88,16 +88,16 @@
 	border-radius: 2px; /* stylelint-disable-line scales/radii */
 }
 
-.p2-join-workspaces__create-workspace-header {
+.p2-join-workspace__create-workspace-header {
 	font-size: 0.875rem;
 	font-weight: 600;
 }
 
-.p2-join-workspaces__create-workspace-subheader {
+.p2-join-workspace__create-workspace-subheader {
 	font-size: 0.75rem;
 }
 
-.p2-join-workspaces__create-workspace-action {
+.p2-join-workspace__create-workspace-action {
 	grid-area: 1 / 2 / span 2;
 	align-items: center;
 	display: flex;
@@ -116,15 +116,15 @@
 	}
 }
 
-.p2-join-workspaces__code-input {
+.p2-join-workspace__code-input {
 	text-align: center;
 }
 
-.p2-join-workspaces__code-input-label {
+.p2-join-workspace__code-input-label {
 	margin-bottom: 1rem;
 }
 
-.p2-join-workspaces__code-input-form {
+.p2-join-workspace__code-input-form {
 	display: flex;
 	flex-direction: column;
 	align-items: center;
@@ -135,25 +135,25 @@
 		margin-bottom: 2rem;
 	}
 
-	button.p2-join-workspaces__code-input-submit,
-	button.p2-join-workspaces__code-input-cancel {
+	button.p2-join-workspace__code-input-submit,
+	button.p2-join-workspace__code-input-cancel {
 		max-width: 60%;
 	}
 
-	.p2-join-workspaces__code-input-cancel {
+	.p2-join-workspace__code-input-cancel {
 		margin-top: 2rem;
 	}
 }
 
-.p2-join-workspaces__code-input-box,
-.p2-join-workspaces__code-input-separator {
+.p2-join-workspace__code-input-box,
+.p2-join-workspace__code-input-separator {
 	height: 4rem;
 	width: 4rem;
 	margin: 0 0.25rem;
 
 }
 
-.p2-join-workspaces__code-input-box {
+.p2-join-workspace__code-input-box {
 	border: 1px solid var( --p2-color-border-medium );
 	font-size: 2rem;
 	text-align: center;
@@ -163,20 +163,20 @@
 	}
 }
 
-.p2-join-workspaces__code-input-separator {
+.p2-join-workspace__code-input-separator {
 	&::before {
 		content: '—';
 	}
 }
 
-.p2-join-workspaces__code-input-error {
+.p2-join-workspace__code-input-error {
 	color: var( --p2-color-error );
 	font-size: 0.875rem;
 	margin-top: 0.5rem;
 }
 
-.p2-join-workspaces__code-input,
-.p2-join-workspaces__workspaces-list {
+.p2-join-workspace__code-input,
+.p2-join-workspace__workspace-list {
 	button.components-button.is-busy {
 		background-image: none;
